requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
python-dateutil>=2.8.0
flask>=2.0.0
